--- frontend/src/pages/PetProfile/PetProfileImageAddition.tsx
@@ -1,5 +1,0 @@
-import PetProfileImageAdditionContent from './PetProfileImageAdditionContent';
-
-const PetProfileImageAddition = () => <PetProfileImageAdditionContent />;
-
-export default PetProfileImageAddition;
