+++ frontend/src/types/petProfile/client.ts
@@ -20,10 +20,9 @@ type PetSize = (typeof PET_SIZES)[number];
 
 type Gender = (typeof GENDERS)[number];
 
-interface PetProfileOutletContextProps {
-  updateIsMixedBreed: (isMixed: boolean) => void;
+interface PetAdditionOutletContextProps {
   updateCurrentStep: (step: number) => void;
   updateIsValidStep: (isValid: boolean) => void;
 }
 
-export type { Breed, Gender, PetProfile, PetProfileOutletContextProps, PetSize };
+export type { Breed, Gender, PetAdditionOutletContextProps, PetProfile, PetSize };
